extenalize entity formatters
